refactor: useSequentialReveal Completions 배열 관리에서 maxIndex 관리로 수정
